feat: use new bottom sheet style in cover plugin
