+++ frontend/appflowy_flutter/lib/mobile/presentation/base/mobile_view_page.dart
@@ -149,7 +149,7 @@ class _MobileViewPageState extends State<MobileViewPage> {
     return AppBarMoreButton(
       onTap: (context) {
         showMobileBottomSheet(
-          context: context,
+          context,
           builder: (_) => _buildViewPageBottomSheet(context),
         );
       },
