+++ frontend/appflowy_flutter/lib/mobile/presentation/bottom_sheet/bottom_sheet.dart
@@ -2,6 +2,7 @@ export 'bottom_sheet_action_widget.dart';
 export 'bottom_sheet_add_new_page.dart';
 export 'bottom_sheet_drag_handler.dart';
 export 'bottom_sheet_rename_widget.dart';
+export 'bottom_sheet_view_item.dart';
 export 'bottom_sheet_view_item_body.dart';
 export 'bottom_sheet_view_item_header.dart';
 export 'bottom_sheet_view_page.dart';
