+++ frontend/appflowy_flutter/lib/mobile/presentation/setting/personal_info/personal_info_setting_group.dart
@@ -45,7 +45,7 @@ class PersonalInfoSettingGroup extends StatelessWidget {
                 trailing: const Icon(Icons.chevron_right),
                 onTap: () {
                   showMobileBottomSheet(
-                    context: context,
+                    context,
                     builder: (_) {
                       return EditUsernameBottomSheet(
                         context,
