+++ frontend/appflowy_flutter/lib/plugins/database_view/widgets/row/cells/date_cell/date_cell.dart
@@ -106,7 +106,7 @@ class _DateCellState extends GridCellState<GridDateCell> {
               text: child,
               onTap: () {
                 showMobileBottomSheet(
-                  context: context,
+                  context,
                   padding: EdgeInsets.zero,
                   builder: (context) {
                     return MobileDateCellEditScreen(
