+++ frontend/appflowy_flutter/lib/plugins/database_view/widgets/row/cells/select_option_cell/select_option_cell.dart
@@ -207,7 +207,7 @@ class _SelectOptionWrapState extends State<SelectOptionWrap> {
         text: child,
         onTap: () {
           showMobileBottomSheet(
-            context: context,
+            context,
             padding: EdgeInsets.zero,
             builder: (context) {
               return MobileSelectOptionEditor(
